move all the functions in separate files
