+++ src/core/quorum/Masternode.js
@@ -1,0 +1,9 @@
+class Masternode {
+  constructor(rpcClient, proTxHash) {
+
+    this.rpc = rpcClient;
+    this.proTxHash = proTxHash;
+  }
+}
+
+module.exports = Masternode;
